lambda/pyproject.toml
[project]
name = "lambda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
edit-img = [
    "loguru>=0.7.3",
    "pillow>=11.2.1",
    "requests>=2.32.3",
]
gen-img = [
    "boto3>=1.37.37",
    "loguru>=0.7.3",
    "openai>=1.75.0",
]
gen-text = [
    "boto3>=1.37.37",
    "loguru>=0.7.3",
    "openai>=1.75.0",
]
pub-img = [
    "loguru>=0.7.3",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "mypy>=1.12.0",
    "pytest>=8.3.3",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
    "ruff>=0.6.9",
]

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
exclude = []

[tool.ruff]
target-version = "py313"
line-length = 88
lint.select = ["ALL"]
lint.ignore = ["D", "B008"]
lint.fixable = ["ALL"]
lint.unfixable = []
